Removing warnings

+++ C++/ProjectEuler/ProjectEuler/Problem006.cpp
@@ -24,7 +24,7 @@ unsigned int squareOfSum(unsigned int naturalNumRange)
 	{
 		sumNum += i;
 	}
-	return pow(sumNum, 2);
+	return static_cast<unsigned int>(pow(sumNum, 2));
 }
 
 
@@ -33,7 +33,7 @@ unsigned int sumOfSquares(unsigned int naturalNumRange)
 	unsigned int sumPow = 0;
 	for (unsigned int i = 0; i < naturalNumRange + 1; ++i) 
 	{
-		sumPow += pow(i, 2);
+		sumPow += static_cast<unsigned int>(pow(i, 2));
 	}
 	return sumPow;
 }
